# public/openapi.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Agent backend service for knowledge capture and recall.",
        "title": "Mindcache FYI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Welcome endpoint",
                "responses": {
                    "200": {
                        "description": "Hello World!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/api/analyse": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyse"
                ],
                "summary": "Analyse a chat and return matching topics and mindcaches",
                "parameters": [
                    {
                        "description": "Chat payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AnalyseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AnalyseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/analyse/clear": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyse"
                ],
                "summary": "Clear cached analysis result for a chat",
                "parameters": [
                    {
                        "description": "Chat payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AnalyseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ClearAnalyseResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mindcache"
                ],
                "summary": "Create a new mindcache from a cached topic",
                "parameters": [
                    {
                        "description": "Topic ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CreateResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/debug/kvcache": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debug"
                ],
                "summary": "View all cached chats and topics (KV debug)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/api/debug/llm": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debug"
                ],
                "summary": "Return LLM provider configuration (no API key)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.LLMConfig"
                        }
                    }
                }
            }
        },
        "/v1/api/delete/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mindcache"
                ],
                "summary": "Delete a mindcache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mindcache ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/api/get/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mindcache"
                ],
                "summary": "Get a single mindcache with its content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mindcache ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mindcache"
                ],
                "summary": "List all mindcaches",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Full-text search over briefs and content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/api/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mindcache"
                ],
                "summary": "Append a cached topic to an existing mindcache",
                "parameters": [
                    {
                        "description": "Mindcache ID and Topic ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UpdateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AnalyseRequest": {
            "type": "object",
            "properties": {
                "chat": {
                    "description": "Chat payload from the extension",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Chat"
                        }
                    ]
                }
            }
        },
        "model.AnalyseResponse": {
            "type": "object",
            "properties": {
                "mindcaches": {
                    "description": "Related (matched) mindcaches",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Mindcache"
                    }
                },
                "topicMindcacheMap": {
                    "description": "Per-topic matched mindcache IDs (topicId → mindcacheId[])",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "topics": {
                    "description": "Extracted topics",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Topic"
                    }
                }
            }
        },
        "model.Chat": {
            "type": "object",
            "properties": {
                "chatId": {
                    "description": "Unique chat identifier",
                    "type": "string"
                },
                "content": {
                    "description": "Full chat content as plain text",
                    "type": "string"
                },
                "messages": {
                    "description": "Structured messages; optional, keeps backward compatibility with flat\ncontent when absent.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "provider": {
                    "description": "Provider enum: gemini, claude, deepseek, yuanbao, chatgpt",
                    "type": "string",
                    "enum": [
                        "gemini",
                        "claude",
                        "deepseek",
                        "yuanbao",
                        "chatgpt"
                    ]
                },
                "sourceUrl": {
                    "description": "Source URL of the conversation",
                    "type": "string"
                },
                "title": {
                    "description": "Chat title",
                    "type": "string"
                }
            }
        },
        "model.ClearAnalyseResponse": {
            "type": "object",
            "properties": {
                "cleared": {
                    "description": "Whether cache entries were cleared",
                    "type": "boolean"
                }
            }
        },
        "model.CreateRequest": {
            "type": "object",
            "properties": {
                "topicId": {
                    "description": "Source topic ID from /analyse response",
                    "type": "string"
                }
            }
        },
        "model.CreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message when success is false",
                    "type": "string"
                },
                "mindcacheId": {
                    "description": "Created mindcache ID",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error description",
                    "type": "string"
                }
            }
        },
        "model.GetResponse": {
            "type": "object",
            "properties": {
                "mainContent": {
                    "description": "Main markdown content",
                    "type": "string"
                },
                "mindcache": {
                    "$ref": "#/definitions/model.Mindcache"
                },
                "sources": {
                    "description": "Provenance records, one per capture that fed this mindcache",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SourceRecord"
                    }
                }
            }
        },
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "mindcaches": {
                    "description": "All mindcaches",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Mindcache"
                    }
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Message text",
                    "type": "string"
                },
                "role": {
                    "description": "Message role: user, assistant, or other",
                    "type": "string"
                }
            }
        },
        "model.Mindcache": {
            "type": "object",
            "properties": {
                "brief": {
                    "description": "Brief summary",
                    "type": "string"
                },
                "id": {
                    "description": "Unique mindcache identifier",
                    "type": "string"
                },
                "sourceUrls": {
                    "description": "Source conversation URLs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "description": "Last update timestamp",
                    "type": "string"
                }
            }
        },
        "model.SearchResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "description": "Ranked matches",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SearchResult"
                    }
                }
            }
        },
        "model.SearchResult": {
            "type": "object",
            "properties": {
                "matchedIn": {
                    "description": "Which indexed fields matched: \"brief\", \"content\", or both",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mindcache": {
                    "description": "Matched mindcache metadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Mindcache"
                        }
                    ]
                },
                "score": {
                    "description": "Relevance score; higher is better",
                    "type": "number"
                },
                "snippet": {
                    "description": "Text window around the first content match; matches are wrapped in\n\\x01/\\x02 marker characters for client-side highlighting",
                    "type": "string"
                }
            }
        },
        "model.SourceRecord": {
            "type": "object",
            "properties": {
                "capturedAt": {
                    "description": "When the record was added",
                    "type": "string"
                },
                "chatId": {
                    "description": "Source chat identifier",
                    "type": "string"
                },
                "excerpts": {
                    "description": "Excerpts of the messages backing the topic",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sourceUrl": {
                    "description": "Source conversation URL",
                    "type": "string"
                },
                "topicBrief": {
                    "description": "Topic brief that was cached",
                    "type": "string"
                },
                "topicTitle": {
                    "description": "Topic title that was cached",
                    "type": "string"
                }
            }
        },
        "model.Topic": {
            "type": "object",
            "properties": {
                "brief": {
                    "description": "Brief summary of the topic",
                    "type": "string"
                },
                "messageRefs": {
                    "description": "MessageRefs are 1-based indices into Chat.Messages supporting this\ntopic. Empty when the chat was captured without structured messages.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sourceChat": {
                    "description": "Source chat ID this topic was extracted from",
                    "type": "string"
                },
                "sourceExcerpts": {
                    "description": "SourceExcerpts are the referenced messages in \"[role] content\" form,\naligned with MessageRefs.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Short topic title",
                    "type": "string"
                },
                "topicId": {
                    "description": "Unique topic identifier",
                    "type": "string"
                }
            }
        },
        "model.UpdateRequest": {
            "type": "object",
            "properties": {
                "mindcacheId": {
                    "description": "Target mindcache ID",
                    "type": "string"
                },
                "topicId": {
                    "description": "Source topic ID from /analyse response",
                    "type": "string"
                }
            }
        },
        "model.UpdateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message when success is false",
                    "type": "string"
                },
                "mindcacheId": {
                    "description": "Target mindcache ID",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.LLMConfig": {
            "type": "object",
            "properties": {
                "baseUrl": {
                    "type": "string"
                },
                "configured": {
                    "type": "boolean"
                },
                "maxConcurrency": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                }
            }
        }
    }
}